requirements.txt
asgiref==3.3.4
boto3==1.34.16
botocore==1.34.16
cloudinary==1.25.0
cryptography==3.4.8
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.3
django-allauth==0.44.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-storages==1.14.2
django-summernote==0.8.11.6
gunicorn==20.1.0
jmespath==1.0.1
oauthlib==3.1.1
Pillow==10.1.0
psycopg2==2.8.6
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
s3transfer==0.10.0
sqlparse==0.4.1
stripe==7.9.0
urllib3==1.26.18
